rename: store -> globalState
