+++ src/globalState.ts
