+++ src/index.ts
@@ -4,7 +4,7 @@
  */
 
 export { loadMicroApp, registerMicroApps, start } from './apis';
-export { initGlobalState } from './store';
+export { initGlobalState } from './globalState';
 export * from './errorHandler';
 export * from './effects';
 export * from './interfaces';
